[77] features 폴더 구성 및 폴더구조 변경
